refactor(ci-scripts): extract prepareReleasePR utils into separate file
