+++ packages/ci-scripts/src/prepareReleasePR/index.ts
@@ -1,0 +1,19 @@
+import {
+  getMergedTodayReleasePRs,
+  setReleasePrWarning,
+  updateReleasePrTitle,
+} from './prepareReleasePR.utils';
+
+export const prepareReleasePR = async ({
+  github,
+  context,
+}: {
+  github: OctoGithub;
+  context: OctoContext;
+}) => {
+  const prsReleasedToday = await getMergedTodayReleasePRs(github, context);
+  const releasedPrsTodayTotal = prsReleasedToday.length;
+  await updateReleasePrTitle(github, context, releasedPrsTodayTotal);
+
+  await setReleasePrWarning(github, context);
+};
